Create test_mcp_scan_validate.py with 31 scan/validation tests
Completed Task features.bees-82b and review fix features.bees-o4v:
- Created test_mcp_scan_validate.py (636 lines) with scan/validation tests
- Extracted 10 test classes: TestGetRepoRoot, TestValidateHivePath,
  TestScanForHive* (Security, ConfigAutoUpdate, BugFixes, FileVsDirectory,
  ExceptionHandling, ErrorPropagation, ConfigHandling, ConfigOptimization)
- Removed scan/validate tests from test_mcp_server.py (~1970 lines removed)
- Fixed duplicate test classes (TestParseHiveFromTicketId removed from test_mcp_server.py)
- Updated README.md with test file organization
- All 1305 tests pass
