+++ .bees/corruption_report.json
@@ -16,5 +16,5 @@
     ],
     "fixes": []
   },
-  "timestamp": "2026-02-05T16:32:45.658702"
+  "timestamp": "2026-02-05T16:43:40.715643"
 }
