+++ docs/tickets/features/features.bees-2i5.md
@@ -32,12 +32,12 @@ description: 'Context: Document the test file split architecture and rationale f
   - Clear rationale for test file organization
 
   - Scan/validate test scope documented'
-parent: features.bees-82b
 up_dependencies:
 - features.bees-le1
-status: open
+parent: features.bees-82b
 created_at: '2026-02-05T16:14:04.955458'
-updated_at: '2026-02-05T16:14:04.955465'
+updated_at: '2026-02-05T16:36:15.082212'
+status: completed
 bees_version: '1.1'
 ---
