+++ docs/tickets/features/features.bees-2sx.md
@@ -42,8 +42,8 @@ down_dependencies:
 - features.bees-qi5
 parent: features.bees-82b
 created_at: '2026-02-05T16:13:49.200001'
-updated_at: '2026-02-05T16:14:11.417642'
-status: open
+updated_at: '2026-02-05T16:35:49.966876'
+status: completed
 bees_version: '1.1'
 ---
