+++ docs/tickets/features/features.bees-5y8.md
@@ -39,8 +39,9 @@ children:
 - features.bees-xab
 - features.bees-se5
 - features.bees-3jb
+- features.bees-o4v
 created_at: '2026-02-05T08:05:49.511857'
-updated_at: '2026-02-05T16:30:29.830552'
+updated_at: '2026-02-05T16:39:04.921894'
 priority: 2
 status: open
 bees_version: '1.1'
