+++ docs/tickets/features/features.bees-61l.md
@@ -34,12 +34,12 @@ description: 'Context: Final verification that all tests pass after extraction a
   - Full test suite passes with 100% success rate
 
   - No test failures or errors'
-parent: features.bees-82b
 up_dependencies:
 - features.bees-qi5
-status: open
+parent: features.bees-82b
 created_at: '2026-02-05T16:14:18.699648'
-updated_at: '2026-02-05T16:14:18.699654'
+updated_at: '2026-02-05T16:37:03.558264'
+status: completed
 bees_version: '1.1'
 ---
