+++ docs/tickets/features/features.bees-82b.md
@@ -34,6 +34,7 @@ up_dependencies:
 - features.bees-4i1
 down_dependencies:
 - features.bees-xab
+- features.bees-o4v
 parent: features.bees-5y8
 children:
 - features.bees-le1
@@ -43,9 +44,9 @@ children:
 - features.bees-qi5
 - features.bees-61l
 created_at: '2026-02-05T16:12:51.341868'
-updated_at: '2026-02-05T16:14:18.703732'
+updated_at: '2026-02-05T16:39:04.926659'
 priority: 0
-status: open
+status: completed
 bees_version: '1.1'
 ---
