+++ docs/tickets/features/features.bees-eun.md
@@ -1,0 +1,58 @@
+---
+id: features.bees-eun
+type: subtask
+title: Add missing TestScanForHiveConfigHandling to test_mcp_scan_validate.py
+description: 'Add TestScanForHiveConfigHandling class to test_mcp_scan_validate.py
+  that was identified as missing during code review.
+
+
+  Context: This test class should have been moved during the test file split but was
+  skipped.
+
+
+  Files to modify:
+
+  - tests/test_mcp_scan_validate.py
+
+
+  Steps:
+
+  1. Find TestScanForHiveConfigHandling in test_mcp_server.py
+
+  2. Move class to test_mcp_scan_validate.py in appropriate location
+
+  3. Ensure imports are correct
+
+  4. Verify test runs successfully
+
+
+  Acceptance:
+
+  - TestScanForHiveConfigHandling class exists in test_mcp_scan_validate.py
+
+  - All tests in the class pass'
+down_dependencies:
+- features.bees-iji
+parent: features.bees-o4v
+created_at: '2026-02-05T16:39:35.528955'
+updated_at: '2026-02-05T16:42:28.870488'
+status: closed
+bees_version: '1.1'
+---
+
+Add TestScanForHiveConfigHandling class to test_mcp_scan_validate.py that was identified as missing during code review.
+
+Context: This test class should have been moved during the test file split but was skipped.
+
+Files to modify:
+- tests/test_mcp_scan_validate.py
+
+Steps:
+1. Find TestScanForHiveConfigHandling in test_mcp_server.py
+2. Move class to test_mcp_scan_validate.py in appropriate location
+3. Ensure imports are correct
+4. Verify test runs successfully
+
+Acceptance:
+- TestScanForHiveConfigHandling class exists in test_mcp_scan_validate.py
+- All tests in the class pass
