+++ docs/tickets/features/features.bees-le1.md
@@ -35,8 +35,8 @@ down_dependencies:
 - features.bees-2i5
 parent: features.bees-82b
 created_at: '2026-02-05T16:13:42.507162'
-updated_at: '2026-02-05T16:14:04.963345'
-status: open
+updated_at: '2026-02-05T16:34:22.111304'
+status: completed
 bees_version: '1.1'
 ---
