+++ docs/tickets/features/features.bees-qi5.md
@@ -41,8 +41,8 @@ down_dependencies:
 - features.bees-61l
 parent: features.bees-82b
 created_at: '2026-02-05T16:14:11.406836'
-updated_at: '2026-02-05T16:14:18.707470'
-status: open
+updated_at: '2026-02-05T16:36:25.011912'
+status: completed
 bees_version: '1.1'
 ---
