+++ docs/tickets/features/features.bees-u8x.md
@@ -1,0 +1,63 @@
+---
+id: features.bees-u8x
+type: subtask
+title: Verify no duplicate tests exist between files
+description: 'Run duplicate detection to ensure no test classes or test methods exist
+  in both test_mcp_server.py and test_mcp_scan_validate.py.
+
+
+  Context: This verification ensures the cleanup is complete and no duplicates remain.
+
+
+  Files to check:
+
+  - tests/test_mcp_server.py
+
+  - tests/test_mcp_scan_validate.py
+
+
+  Steps:
+
+  1. Extract all test class names from both files
+
+  2. Extract all test method names from both files
+
+  3. Check for overlaps/duplicates
+
+  4. Report any duplicates found
+
+
+  Acceptance:
+
+  - No test classes appear in both files
+
+  - No test methods appear in both files
+
+  - Report confirms zero duplicates'
+down_dependencies:
+- features.bees-iji
+parent: features.bees-o4v
+created_at: '2026-02-05T16:39:38.932501'
+updated_at: '2026-02-05T16:43:11.763985'
+status: closed
+bees_version: '1.1'
+---
+
+Run duplicate detection to ensure no test classes or test methods exist in both test_mcp_server.py and test_mcp_scan_validate.py.
+
+Context: This verification ensures the cleanup is complete and no duplicates remain.
+
+Files to check:
+- tests/test_mcp_server.py
+- tests/test_mcp_scan_validate.py
+
+Steps:
+1. Extract all test class names from both files
+2. Extract all test method names from both files
+3. Check for overlaps/duplicates
+4. Report any duplicates found
+
+Acceptance:
+- No test classes appear in both files
+- No test methods appear in both files
+- Report confirms zero duplicates
